requirements.txt
streamlit
pandas
scikit-learn
umap-learn
plotly
